rename: API.js 파일 src/api 폴더 생성후 이동
